+++ src/app/(auth)/reset-password/page.tsx
@@ -1,70 +1,13 @@
 'use client';
-import { Button } from '@/components/ui/button';
-import { z } from 'zod';
-import { useForm } from 'react-hook-form';
-import { zodResolver } from '@hookform/resolvers/zod';
-import { Form } from '@/components/ui/form';
-import { useState } from 'react';
-import Link from 'next/link';
-import { InputTextField } from '@/components/molecules/input-text-field';
 
-const formSchema = z.object({
-	email: z.string().nonempty({ message: '이메일을 입력해주세요' }).email({
-		message: '유효한 이메일이 아닙니다.',
-	}),
-});
+import Link from 'next/link';
+import FindEmailForm from '@/components/organisms/find-email-form';
 
 export default function Page() {
-	const form = useForm<z.infer<typeof formSchema>>({
-		resolver: zodResolver(formSchema),
-		defaultValues: {
-			email: '',
-		},
-	});
-
-	const [isSuccessEmailSend, setIsSuccessEmailSend] = useState(false);
-	const [isNotExisted, setIsNotExisted] = useState(false);
-
-	// 비밀번호 리셋
-	const onSubmit = async (values: z.infer<typeof formSchema>) => {
-		console.log(values);
-		try {
-			// TODO: 비밀번호 찾기 로직 작성 /find-email
-			setIsSuccessEmailSend(true);
-		} catch (e) {
-			// TODO: 비밀번호 찾기 실패시 에러코드 맞춰서 설정해주기
-			setIsNotExisted(true);
-			console.log(e);
-		}
-	};
-
-	return (
-		<div>
-			<Form {...form}>
-				<form onSubmit={form.handleSubmit(onSubmit)} className="space-y-8">
-					<InputTextField
-						form={form}
-						name="email"
-						label="이메일"
-						type="email"
-						placeholder="myemail@mail.com"
-					/>
-
-					{/* 성공시 로그인페이지로 갈지 비밀번호 찾기로 갈지 물어보기 */}
-					{isSuccessEmailSend && (
-						<div className="flex flex-col">
-							<p className="text-blue-900">이메일을 확인해주세요!</p>
-							<Link href="/login">로그인하러 가기</Link>
-						</div>
-					)}
-
-					{isNotExisted && (
-						<p className="text-red-600">해당 이메일이 존재하지 않습니다</p>
-					)}
-					<Button>이메일 찾기</Button>
-				</form>
-			</Form>
-			<Link href="/login">로그인으로 돌아가기</Link>
-		</div>
-	);
+  return (
+    <div>
+      <FindEmailForm isResetPassword />
+      <Link href="/login">로그인으로 돌아가기</Link>
+    </div>
+  );
 }
